% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_go_groups.R
\name{add_clusters_to_go}
\alias{add_clusters_to_go}
\title{Add simplifyEnrichment semantic similarity clustering to GO ids}
\usage{
add_clusters_to_go(goids, ...)
}
\arguments{
\item{goids}{vector of GO ids to cluster}

\item{...}{arguments to be passed to simplifyEnrichment::GO_similarity and simplifyEnrichment::simplifyGO}
}
\value{
data.frame with GO ids and clusters
}
\description{
Add simplifyEnrichment semantic similarity clustering to GO ids
}
\examples{
enrich_file = system.file("extdata","go_enrich_results.txt", package = "goreparent")
go_input = read.table(enrich_file, sep="\t", header=TRUE)
clusters = add_clusters_to_go(go_input$ID)
}
