% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_go_groups.R
\name{add_go_groups}
\alias{add_go_groups}
\title{Annotate GO terms into groups of related terms}
\usage{
add_go_groups(
  go_input,
  FDR = 0.05,
  n_top = 10,
  descriptive_parent = TRUE,
  max_parents = 2,
  max_from_top = 3,
  ignore_terms = NULL,
  max_children = NULL,
  method = c("parent", "cluster"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{go_input}{list of or single object (\code{data.frame}, \code{enrichResult},\code{vector}) containing GO ids.}

\item{FDR}{minimum FDR/adjusted p-value required for grouping of GO terms. default = \code{0.05}}

\item{n_top}{number of groups to add descriptive parents to.
This cuts down on processing time and only annotates the groups with the n-th highest minimum FDR.
default = \code{10}}

\item{descriptive_parent}{add a more descriptive annotation of groups based on common relatives?}

\item{max_parents}{if using method = \code{"parent"}:
how many levels up from a GO term should be used to find common parents?
default = \code{2}}

\item{max_from_top}{if using method = \code{"parent"}:
how many levels down from top level GO terms should GO terms be excluded when searching for parent terms?
default = \code{3}
Using this option prevents grouping of GO terms into top levels terms e.g. level 1: "biological_process" level 2: "biological regulation", "behavior"}

\item{ignore_terms}{if using method = \code{"parent"}:
vector of any GO ids to ignore when searching for common parents.
Can be helpful if you need to force one group into two or more smaller, more specific groups.
default = \code{NULL}}

\item{max_children}{if using method = \code{"parent"}:
ignore all GO parent terms with at least this many 'children' terms.
default = \code{NULL}}

\item{method}{which method to use for grouping.
\code{"parent"}: common parents of GO terms, going up two levels and ignoring any top level terms/terms on the ignore list.
\code{"cluster"}: simplifyEnrichment semantic similarity clustering.
Note that these terms may not have very direct parent/child relations, and non-descriptive group names,
therefore running with \code{descriptive_parent = TRUE} is highly recommended.}

\item{verbose}{run with status messages? set to \code{FALSE} to supress messages.}

\item{...}{arguments to be passed to \code{add_clusters_to_go}}
}
\value{
data.frame with original GO terms, new groups, and parent terms.
Contains the same columns as any data.frame (or object able to be conevrted to a data.frame) passed in as \code{go_input}.
}
\description{
Takes enrichGO, chipenrich, or vectors of GO terms and groups them into related terms.
Either using "parent" grouping, which finds common parents of terms,
or using "cluster" grouping relying on simplifyEnrichment semantic similarity clustering.
Additionally optionally adds additional group descriptions based on most common related terms
(using both direct parent-child relationships and uncle/cousin relationships)
overing all GO terms within the group.
}
\examples{

enrich_file = system.file("extdata","go_enrich_results.txt", package = "goreparent")
go_input = read.table(enrich_file, sep="\t", header=TRUE)
go_parents = add_go_groups(go_input, descriptive_parent=FALSE)
go_clusters = add_go_groups(go_input, descriptive_parent=FALSE, method = "cluster")

}
