% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_go_ratios}
\alias{calc_go_ratios}
\title{Recalculate GO ratios from enrichResults to a numeric variable}
\usage{
calc_go_ratios(go_results)
}
\description{
Recalculate GO ratios from enrichResults to a numeric variable
}
\keyword{internal}
