% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_go_datatype}
\alias{check_go_datatype}
\title{Check format/datatype of input GO ids}
\usage{
check_go_datatype(input)
}
\description{
Checks if input is an enrichResult (clusterProfiler), or other object which can be converted to a data.frame
Checks colname of column which contains GO ids
}
\keyword{internal}
