% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_goid_col}
\alias{check_goid_col}
\title{Check which column name is used for GO ids}
\usage{
check_goid_col(go_results, verbose = TRUE)
}
\description{
Check which column name is used for GO ids
}
\keyword{internal}
