% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_signif_col}
\alias{check_signif_col}
\title{Check which column name is used for FDR/pvalues from a list of common terms}
\usage{
check_signif_col(
  go_results,
  signif_col = c("FDR", "p.adjust", "pval", "p", "p.val")
)
}
\description{
Checks for FDRs/ adjusted pvalues FIRST
}
\keyword{internal}
