% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_go_parents_utils.R
\name{get_direct_go_path}
\alias{get_direct_go_path}
\title{get the direct path between two GO terms with any kind of offspring/ancestor relationship}
\usage{
get_direct_go_path(child, parent, top_level_terms = NULL)
}
\arguments{
\item{child}{child GO id}

\item{parent}{parent GO id}

\item{top_level_terms}{any 'top level terms' to remove from analysis}
}
\description{
get the direct path between two GO terms with any kind of offspring/ancestor relationship
}
\examples{

get_direct_go_path(child = "GO:0021955", parent = "GO:0061564")
# returns ALL potential direct pathways
get_direct_go_path(child = "GO:0021955", parent = "GO:0031175")
# returns empty if parent:child relationship does not exist
get_direct_go_path(child = "GO:0021955", parent = "GO:0050890")
}
