% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_go_parents.R
\name{get_meta_parents}
\alias{get_meta_parents}
\title{from a list of all two-child->parent relationships, merge parent terms into 'Meta-parents'
that cover as many of the parent terms as possible that have direct parent:child relationships with each other}
\usage{
get_meta_parents(go_parents)
}
\arguments{
\item{go_parents}{data.frame with two go terms and their corresponding 'parent'}
}
\description{
from a list of all two-child->parent relationships, merge parent terms into 'Meta-parents'
that cover as many of the parent terms as possible that have direct parent:child relationships with each other
}
\keyword{internal}
