% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_go_parents_utils.R
\name{get_stepchildren}
\alias{get_stepchildren}
\title{Get "step-child" terms for GO terms}
\usage{
get_stepchildren(
  go_terms,
  return_partner = FALSE,
  return_descriptions = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{go_terms}{vector of GO ids}
}
\description{
stepchild are found by finding partners (now birth partners) (GO -> all child terms (1 level down) -> all parent terms (1 level up))
then all offspring of partner terms (all offspring of birth partners -> step children) (from above: all parent terms (1 level up)/"partners" -> all child terms)
}
\keyword{internal}
