% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_top_level_goterms}
\alias{get_top_level_goterms}
\title{Makes a vector containing "top level" GO terms,}
\usage{
get_top_level_goterms(max_parents = 2, max_from_top = 2, ignore_terms = NULL)
}
\description{
Used to exclude when generating parent terms as ALL terms will be under the umbrella of
"BP/MF/CC" etc and this creates groups which are too large and non-specific
}
\keyword{internal}
