% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_top_parent_descriptions}
\alias{get_top_parent_descriptions}
\title{Sort a GO_result with groups (in parent_description) by "significance"
and return the top n parent terms
Used to prevent plotting/further analysis of groups which are less significant.}
\usage{
get_top_parent_descriptions(
  go_results,
  n_top = 10,
  group_col = "parent_description",
  ...
)
}
\description{
Sort a GO_result with groups (in parent_description) by "significance"
and return the top n parent terms
Used to prevent plotting/further analysis of groups which are less significant.
}
\keyword{internal}
