% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_go_parents_utils.R
\name{get_top_parents_df}
\alias{get_top_parents_df}
\title{Using defined lists of parents for go terms, find parent terms that cover all go terms}
\usage{
get_top_parents_df(
  all_go_terms,
  parents,
  children_new,
  stepchild_new,
  stepchild_ratio = T
)
}
\arguments{
\item{stepchild_ratio}{use total STEPCHILDREN (\code{TRUE}) or total CHILDREN (\code{FALSE}) of parent terms to calculate ratios/specificity of stepparent terms
Using total CHILDREN (\code{FALSE}) may result in ratios higher than 1.}
}
\description{
using the stepchild relationships
AND order by specificity of parent/stepparent terms on the defined go terms
}
\keyword{internal}
