% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id2term}
\alias{id2term}
\title{Convert GO ids to GO descriptions}
\usage{
id2term(id, replace = T)
}
\arguments{
\item{id}{character vector with GO ids in the format "GO:XXXXXXX"}

\item{replace}{replace any non-GO terms with NA \code{TRUE} or return original values \code{FALSE}}
}
\value{
character vector with GO descriptions
}
\description{
Convert GO ids to GO descriptions
}
\examples{
id2term("GO:0050804")
id2term(c("GO:0050804","GO:0099177"))
id2term(c("GO:0050804", NA, "GO:0099177"))
}
