% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_go_parents.R
\name{lowest_common_parent}
\alias{lowest_common_parent}
\title{find the lowest level term that is a direct parent of (or is) two GO terms.
Runs for all potential combinations of GO terms given as input in \code{go_term_list}}
\usage{
lowest_common_parent(
  go_term_list,
  max_parents = 2,
  max_from_top = 2,
  fill_terms = T,
  ignore_terms = NULL
)
}
\arguments{
\item{fill_terms}{fill in parent terms using parent terms that have already been established in the first use of the parent matching?}
}
\description{
find the lowest level term that is a direct parent of (or is) two GO terms.
Runs for all potential combinations of GO terms given as input in \code{go_term_list}
}
\keyword{internal}
