% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_parents.R
\name{plot_go_parents}
\alias{plot_go_parents}
\title{plot the output from a goreparent add_go_group()}
\usage{
plot_go_parents(
  go_results,
  type = "dot",
  collapse = TRUE,
  n_top = 10,
  sort_top_by = c("FDR", "p.adjust", "pval", "p", "p.val"),
  color = "-log10(FDR)",
  size = NULL,
  x.axis = c("GeneRatio", "Odds.Ratio", "-log10(FDR)")
)
}
\arguments{
\item{go_results}{output from add_go_groups()
data.frame with original GO terms, new groups, and parent terms.
must have column named "parent_description"
and optionally "Description" for displaying individual GO ids (with \code{collapse=FALSE})}

\item{type}{\code{"dot"}. Currently nonfunctional, only dot available.}

\item{collapse}{collapse individual terms onto the same y-axis value for each group?. default \code{TRUE}}

\item{n_top}{number of groups to plot, ordered by \code{sort_top_by}.
default = \code{10}}

\item{sort_top_by}{how to sort groups for n_top. The first match to a character variable will be used.
default \code{c("FDR", "p.adjust", "pval", "p", "p.val")}}

\item{color}{which variable to color dots by. default \code{"FDR"}}

\item{size}{which variable to size dots by. default \code{NULL}}

\item{x.axis}{which variable to use as the x-axis. default \code{c("GeneRatio", "Odds.Ratio")}}
}
\description{
plot the output from a goreparent add_go_group()
}
\examples{
enrich_file = system.file("extdata","go_enrich_results.txt", package = "goreparent")
go_input = read.table(enrich_file, sep="\t", header=TRUE)
go_parents = add_go_groups(go_input, descriptive_parent=FALSE, max_from_top=3)
go_clusters = add_go_groups(go_input, descriptive_parent=TRUE, method = "cluster")

plot_go_parents(go_parents)
plot_go_parents(go_clusters)

# un-collapse parent terms and show all children for the top 2 categories
plot_go_parents(go_parents, collapse=FALSE, n_top=2)

}
