% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_split_vector}
\alias{str_split_vector}
\title{apply str_split to string vectors}
\usage{
str_split_vector(string_vector, split_by, position)
}
\arguments{
\item{string_vector}{vector of strings}

\item{split_by}{character delimiter to split strings by}

\item{position}{which position of the split to return}
}
\value{
vector of strings split by the split_by character at the position
}
\description{
apply str_split to string vectors
}
\examples{
str_split_vector(c("ABC_DEF_HIJ_KLM", "NOP_QRS_TUV_WXY"), "_" , 2)

str_split_vector(c("ABC_DEF_HIJ_KLM", "NOP_QRS_TUV"), "_" , 2)

str_split_vector(c("ABC_DEF_HIJ_KLM", "NOP_QRS_TUV"), "_" , -1)
}
\author{
Brandon Signal
}
